products/pyproject.toml
[tool.poetry]
name = "products"
version = "0.1.0"
description = ""
authors = ["Strawberry"]

[tool.poetry.dependencies]
python = "^3.7"
strawberry-graphql = { path = "../../../"}
starlette = "^0.14.2"

[tool.poetry.dev-dependencies]
uvicorn = "^0.13.4"
taskipy = "^1.8.1"

[tool.taskipy.tasks]
server = "uvicorn app:app --port 7200 --reload"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
